pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orkl-mcp-server"
version = "0.1.0"
description = "An MCP Server that interfaces with the ORKL Threat Intelligence API"
readme = "README.md"
authors = [
    { name = "Heith Seewald", email = "heiths@gmail.com" }
]

license = { text = "MIT" }
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["threat-intelligence", "mcp", "api", "security", "orkl"]
dependencies = [
    "fastmcp>=0.1.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.urls]
Homepage = "https://github.com/heiths/orkl_mcp_server"
Documentation = "https://github.com/heiths/orkl_mcp_server#readme"
Repository = "https://github.com/heiths/orkl_mcp_server.git"
Issues = "https://github.com/heiths/orkl_mcp_server/issues"

[project.scripts]
orkl-mcp-server = "run_server:run_server"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.10.0",
    "pytest-xdist>=3.3.0",
    "anyio>=3.7.0",
    "ruff>=0.0.249",
    "pyright>=1.1.316",
    "mock>=5.0.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.3",
]


[tool.ruff]
target-version = "py310"
ignore = ["E501"]
extend-exclude = ["docs", "tests"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
addopts = "--ignore=tests/test_integration.py"
markers = [
    "integration: marks tests as integration tests"
]

[tool.pyright]
include = ["orkl_mcp"]
typeCheckingMode = "basic"
pythonVersion = "3.10"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = [".git", ".venv", "docs", "tests"]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.26.0",
]
